# api/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0-oas3
  title: hived
  description: hived's API
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/xashmith/hived/0.1'
paths:
  /price:
    get:
      description: Returns the price of the crypto
      parameters:
        - name: name
          in: query
          description: the symbol of the cryptocurrency
          schema:
            type: string
        - name: unit
          in: query
          description: the unit the return the price in
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                  - unit
                  - price
                properties:
                  name:
                    type: string
                  unit:
                    type: string
                  price:
                    type: number
  /pair:
    get:
      description: Returns the ratio of one to two multiplied by a value
      parameters:
        - name: one
          in: query
          description: the name of the currency that's going to be multiplied
          schema:
            type: string
        - name: two
          in: query
          description: the name of the second currency
          schema:
            type: string
        - name: multiplier
          in: query
          description: the amount hte first currency is going to be multiplied
          schema:
            type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                required:
                  - ratio
                properties:
                  ratio:
                    type: number
  /addalerts:
    post:
      description: Add alerts to the alertmanager's list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - expr
              properties:
                name:
                  type: string
                expr:
                  type: string
      responses:
        '200':
          description: successful update
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
                  isSuccessful:
                    type: boolean